Remove old password from password reset, require it on password update.
Typically people want to reset it because they can't remember it, so
remove the requirement for the old password. Also fix a few issues when
submitting the form, the url and the old==new check.

When updating the users info, if a new pass is given check if the old
password is correct.